+++ src/components/collection-item/collection-item.js
