+++ src/components/collection-item/collection-item.styles.scss
