+++ src/components/collection-preview/collection-preview.js
