+++ src/components/collection-preview/collection-preview.styles.scss
